refacor: 축제 등록 수정 공통 컴포넌트 사용 분리
