+++ src/components/festivalForm/FestivalFormAddressCard.tsx
@@ -2,14 +2,14 @@ import BorderCardComponent from '@/components/common/BorderCardComponent';
 import type { FestivalCardProps } from '@/types/FestivalFormTypes';
 
 /**
- * 축제 등록 주소 카드
+ * 축제 등록,수정 폼 주소 카드
  * @param formData - 폼 데이터
  * @param handleInputChange - 입력 변경 핸들러
  * @param isSubmitting - 제출 중 여부
- * @returns 축제 등록 주소 카드 컴포넌트
- * 축제 등록 주소를 표시합니다.
+ * @returns 축제 등록,수정 폼 주소 카드 컴포넌트
+ * 축제 등록,수정 폼 주소를 표시합니다.
  */
-const RegisterFestivalAddressCard = ({
+const FestivalFormAddressCard = ({
   formData,
   handleInputChange,
   isSubmitting,
@@ -45,4 +45,4 @@ const RegisterFestivalAddressCard = ({
   );
 };
 
-export default RegisterFestivalAddressCard;
+export default FestivalFormAddressCard;
