+++ src/components/festivalForm/FestivalFormAreaCard.tsx
@@ -1,16 +1,16 @@
 import BorderCardComponent from '@/components/common/BorderCardComponent';
 import AREA_OPTIONS from '@/utils/areaOptions';
-import type { FestivalCardProps } from '../../../types/FestivalFormTypes';
+import type { FestivalCardProps } from '@/types/FestivalFormTypes';
 
 /**
- * 축제 등록 지역 카드
+ * 축제 등록,수정 폼 지역 카드
  * @param formData - 폼 데이터
  * @param handleInputChange - 입력 변경 핸들러
  * @param isSubmitting - 제출 중 여부
- * @returns 축제 등록 지역 카드 컴포넌트
- * 축제 등록 지역을 표시합니다.
+ * @returns 축제 등록,수정 폼 지역 카드 컴포넌트
+ * 축제 등록,수정 폼 지역을 표시합니다.
  */
-const RegisterFestivalAreaCard = ({
+const FestivalFormAreaCard = ({
   formData,
   handleInputChange,
   isSubmitting,
@@ -43,4 +43,4 @@ const RegisterFestivalAreaCard = ({
   );
 };
 
-export default RegisterFestivalAreaCard;
+export default FestivalFormAreaCard;
