+++ src/components/festivalForm/FestivalFormOverviewCard.tsx
@@ -1,15 +1,15 @@
 import BorderCardComponent from '@/components/common/BorderCardComponent';
-import type { FestivalCardProps } from '../../../types/FestivalFormTypes';
+import type { FestivalCardProps } from '@/types/FestivalFormTypes';
 import TextInputWithCounter from '@/components/form/TextInputWithCounter';
 /**
- * 축제 등록 개요 카드
+ * 축제 등록,수정 폼 개요 카드
  * @param formData - 폼 데이터
  * @param handleInputChange - 입력 변경 핸들러
  * @param isSubmitting - 제출 중 여부
- * @returns 축제 등록 개요 카드 컴포넌트
- * 축제 등록 개요를 표시합니다.
+ * @returns 축제 등록,수정 폼 개요 카드 컴포넌트
+ * 축제 등록,수정 폼 개요를 표시합니다.
  */
-const RegisterFestivalOverviewCard = ({
+const FestivalFormOverviewCard = ({
   formData,
   handleInputChange,
   isSubmitting,
@@ -39,4 +39,4 @@ const RegisterFestivalOverviewCard = ({
   );
 };
 
-export default RegisterFestivalOverviewCard;
+export default FestivalFormOverviewCard;
