+++ src/components/festivalForm/FestivalFormPeriodCard.tsx
@@ -1,15 +1,15 @@
 import BorderCardComponent from '@/components/common/BorderCardComponent';
-import type { FestivalCardProps } from '../../../types/FestivalFormTypes';
+import type { FestivalCardProps } from '@/types/FestivalFormTypes';
 
 /**
- * 축제 등록 기간 카드
+ * 축제 등록,수정 폼 기간 카드
  * @param formData - 폼 데이터
  * @param handleInputChange - 입력 변경 핸들러
  * @param isSubmitting - 제출 중 여부
- * @returns 축제 등록 기간 카드 컴포넌트
- * 축제 등록 기간을 표시합니다.
+ * @returns 축제 등록,수정 폼 기간 카드 컴포넌트
+ * 축제 등록,수정 폼 기간을 표시합니다.
  */
-const RegisterFestivalPeriodCard = ({
+const FestivalFormPeriodCard = ({
   formData,
   handleInputChange,
   isSubmitting,
@@ -46,4 +46,4 @@ const RegisterFestivalPeriodCard = ({
   );
 };
 
-export default RegisterFestivalPeriodCard;
+export default FestivalFormPeriodCard;
